% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barddt.R
\name{barddt}
\alias{barddt}
\title{Main function to run BARDDT model}
\usage{
barddt(
  y,
  x,
  w,
  c,
  delta = 0.1,
  num_chains = 20,
  num_gfr = 2,
  num_burnin = 0,
  num_mcmc = 500,
  parallel = FALSE,
  ncores = 10
)
}
\arguments{
\item{y}{Output variable}

\item{x}{Running variable}

\item{w}{Additional variables}

\item{c}{Cutoff value}

\item{delta}{Test set window}

\item{num_chains}{Number of parallel runs}

\item{num_gfr}{GFR samples}

\item{num_burnin}{Burnin}

\item{num_mcmc}{MCMC samples per chain}

\item{parallel}{bool: TRUE if running in parallel}

\item{ncores}{number of cores used for parallel computing}
}
\value{
Posterior samples for the RDD CATE
}
\description{
Main function to run BARDDT model
}
\examples{
example
}
